Move preference stuff into one file
